requirements.txt
# Core dependencies
fastapi==0.109.0
httpx==0.26.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
typer==0.9.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
black==23.12.1
ruff==0.1.11
mypy==1.8.0
isort==5.13.2

# Type stubs
types-PyYAML==6.0.12.12